Implement code for 2/3 power ups and set up code for the 3rd one
Implemented the double points power up, the eliminate wrong answer power up and parts of the halfTime power up. In order to finish the last one, first we need to fix the bug regarding the display of the template question instead of starting a game. Then we need to alter the server and client side timers. After that we need to alter the player points award system.
Note: the code is not tested, since another bug prevents it from working.