Fix bug with gameId being joined.
The gameId didn't use to get updated until joinGame is called, therefore checkUsername checked for the username against the last game played (or gameId=0 for the initial run). The new setup methods are used to separate the gameId generation from joinGame so they can be called before checkUsername is called.