Forbid user from submitting an answer after their time has ran out
Forbid user from submitting an answer after their time has ran out in the case that someone else used a halfTime power up. This was not an unfairness issue, but if the user sends an answer with -2 seconds left, they will lose points. To avoid the situation, this commit is implemented.